requirements.txt
backcall==0.2.0
decorator==5.0.9
Django==2.2
django-extensions==3.1.3
django-location-field==2.1.0
geographiclib==1.50
geopy==2.1.0
ipdb==0.13.9
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.18
psycopg2==2.8
ptyprocess==0.7.0
Pygments==2.9.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==4.3.3
wcwidth==0.2.5
